将计算功能模块化
